Update mysql_zencart.sql
I think it would be easier for the administrator to understand the maximum values under configuration if they were sorted differently.  
I have changed the sort order for the items that fall within administrator/admin area so that the "storefront" items are at the top, then administrator functions, then miscellaneous.  It may just be an "OCD" thing... but I really think it will help to have them sorted differently.  Thanks for your consideration of this change.